cleanup eventwatcher files
